Crate tof_dataclasses

Tof dataclasses

This crate provides tof related dataclasses for

events

network i/o wrappers (“packets”) for classes

function related constants

calibration

commands/responses

TODO: alerts

Modules

analysis

Implementations of analysis engine This is based on the original code by J.Zweerink

calibrations

The ReadoutBoardCalibration is the first stage in a 3 stage calibration process for the tof

commands

Commmands which can be issued to the various components of the tof system.

constants

Global constants for tof operations

errors

Specific error types

events

Events

io

Input/Output

manifest

Entities of the tof, that is LocalTriggerBoard, ReadoutBoard, PB, MTB…

monitoring

Structures for monitoring

packets

Packets are a way to send data over the network.

run

serialization

Serialization/Deserialization helpers

threading

Structs

BitMask32

Representation of 32 bit mask

Traits

FromRandom

Create structures filled with random number to be used for testing and benchmarking

Type Definitions

DsiLtbRBMapping

A type for the master trigger mappings