Struct tof_dataclasses::calibrations::RBCalibrations

pub struct RBCalibrations {
    pub rb_id: u8,
    pub d_v: f32,
    pub timestamp: u32,
    pub serialize_event_data: bool,
    pub v_offsets: [[f32; 1024]; 9],
    pub v_dips: [[f32; 1024]; 9],
    pub v_inc: [[f32; 1024]; 9],
    pub tbin: [[f32; 1024]; 9],
    pub vcal_data: Vec<RBEvent>,
    pub tcal_data: Vec<RBEvent>,
    pub noi_data: Vec<RBEvent>,
}

Fields
rb_id: u8d_v: f32timestamp: u32serialize_event_data: boolv_offsets: [[f32; 1024]; 9]v_dips: [[f32; 1024]; 9]v_inc: [[f32; 1024]; 9]tbin: [[f32; 1024]; 9]vcal_data: Vec<RBEvent>tcal_data: Vec<RBEvent>noi_data: Vec<RBEvent>
Implementations

impl RBCalibrations

pub const NSKIP: usize = 2usize

pub const SINMAX: usize = 60usize

pub const DVCUT: f32 = 15f32

pub const NOMINALFREQ: f32 = 2f32

pub const CALFREQ: f32 = 0.0250000004f32

pub fn clean_input_data(&mut self)

Remove events with invalid traces or event fragment bits set

pub fn apply_vcal_constants( &self, adc: &Vec<f32>, channel: usize, stop_cell: usize ) -> Vec<f32>

pub fn timing_calibration( &self, edge: &Edge, apply_vcal: bool ) -> Result<Vec<Vec<f32>>, CalibrationError>

Voltage calibration has to be applied

Returns

vec[ch[9], tbin[1024]]

pub fn calibrate(&mut self) -> Result<(), CalibrationError>

Call to the calibration routine, using the set input data

pub fn spike_cleaning( voltages: &mut Vec<Vec<f32>>, stop_cell: u16 ) -> Result<(), WaveformError>

Apply the spike cleaning to all channels

pub fn voltages( &self, channel: usize, stop_cell: usize, adc: &Vec<u16>, waveform: &mut Vec<f32> )

Apply the voltage calibration to a single channel FIXME - mixing of naming conventions for the channels

FIXME - make it return Result<(), CalibrationError>

Arguments

channel : Channel id 1-9

stop_cell : This channels stop cell

adc : Uncalibrated channel data

waveform : Pre-allocated array to hold calibrated waveform data.

pub fn nanoseconds( &self, channel: usize, stop_cell: usize, times: &mut Vec<f32> )

Apply the timing calibration to a single channel

This will allocate the array for the waveform time bins (unit is ns)

Arguments

channel : Channel id 1-9

stop_cell : This channels stop cell

pub fn new(rb_id: u8) -> Self

pub fn discard_data(&mut self)

Discard the data to reduce the memory footprint

pub fn from_file( filename: String, discard_data: bool ) -> Result<Self, SerializationError>

Gets the calibration from a file which has the RBCalibration stored in a TofPacket

E.g. if it was written with TofPacketWriter

pub fn from_txtfile(filename: &Path) -> Self

pub fn get_id_from_filename(&mut self, filename: &Path) -> u8

Infer the readoutboard id from the filename

Assuming a certain naming scheme for the filename “rbXX_cal.txt” we extract the readoutboard id