tof_dataclasses::commands
Module config

Payloads for commands that configure an entity of the TOF system.

Structs

AnalysisEngineConfig

Analysis Engine Config Settings to change the configuration of the analysis engine (pulse extraction)

DataPublisherConfig

LTBThresholdConfig

Set ltb thresholds

PreampBiasConfig

Set preamp voltages

RBChannelMaskConfig

RunConfig

Readoutboard configuration for a specific run

TOFEventBuilderConfig

TOF Event Builder Settings Configuring the TOF event builder during flight If a setting is set to None, it will keep the previous setting

TofRBConfig

TofRunConfig

TriggerConfig

Enums

BuildStrategy

Build Strategy