tof_dataclasses::commands::config
Struct TOFEventBuilderConfig

pub struct TOFEventBuilderConfig {

    pub active_fields: u32,
    pub cachesize: Option<u32>,
    pub n_mte_per_loop: Option<u32>,
    pub n_rbe_per_loop: Option<u32>,
    pub te_timeout_sec: Option<u32>,
    pub sort_events: Option<bool>,
    pub build_strategy: Option<BuildStrategy>,
    pub wait_nrb: Option<u8>,
    pub greediness: Option<u8>,
    pub hb_send_interval: Option<u16>,
    pub only_save_interesting: Option<bool>,
    pub thr_n_hits_umb: Option<u8>,
    pub thr_n_hits_cbe: Option<u8>,
    pub thr_n_hits_cor: Option<u8>,
    pub thr_tot_edep_umb: Option<f32>,
    pub thr_tot_edep_cbe: Option<f32>,
    pub thr_tot_edep_cor: Option<f32>,

}

TOF Event Builder Settings Configuring the TOF event builder during flight If a setting is set to None, it will keep the previous setting

Fields
active_fields: u32cachesize: Option<u32>n_mte_per_loop: Option<u32>n_rbe_per_loop: Option<u32>te_timeout_sec: Option<u32>sort_events: Option<bool>build_strategy: Option<BuildStrategy>wait_nrb: Option<u8>greediness: Option<u8>hb_send_interval: Option<u16>only_save_interesting: Option<bool>thr_n_hits_umb: Option<u8>thr_n_hits_cbe: Option<u8>thr_n_hits_cor: Option<u8>thr_tot_edep_umb: Option<f32>thr_tot_edep_cbe: Option<f32>thr_tot_edep_cor: Option<f32>
Implementations

impl TOFEventBuilderConfig

pub fn new() -> Self

pub fn set_cachesize(&mut self, csize: u32)

pub fn set_n_mte_per_loop(&mut self, n: u32)

pub fn set_n_rbe_per_loop(&mut self, n: u32)

pub fn set_te_timeout_sec(&mut self, te: u32)

pub fn set_sort_events(&mut self, sort: bool)

pub fn set_build_strategy(&mut self, bs: BuildStrategy)

pub fn set_wait_nrb(&mut self, nrb: u8)

pub fn set_greediness(&mut self, greed: u8)

pub fn set_hb_send_interval(&mut self, interval: u16)

pub fn set_only_save_interesting(&mut self, do_it: bool)

pub fn thr_n_hits_umb(&mut self, nhit: u8)

pub fn thr_n_hits_cbe(&mut self, nhit: u8)

pub fn thr_n_hits_cor(&mut self, nhit: u8)

pub fn thr_tot_edep_umb(&mut self, thr: f32)

pub fn thr_tot_edep_cbe(&mut self, thr: f32)

pub fn thr_tot_edep_cor(&mut self, thr: f32)