Enum tof_dataclasses::commands::TofCommand

pub enum TofCommand {

    Unknown(u32),
    Ping(u32),
    Moni(u32),
    Power(u32),
    SetThresholds(u32),
    SetMTConfig(u32),
    SetPreampBias(u32),
    DataRunStop(u32),
    DataRunStart(u32),
    StartValidationRun(u32),
    GetFullWaveforms(u32),
    NoiCalibration(u32),
    VoltageCalibration(u32),
    TimingCalibration(u32),
    DefaultCalibration(u32),
    UnspoolEventCache(u32),
    SetRBDataBufSize(u32),
    TriggerModeForced(u32),
    TriggerModeForcedMTB(u32),
    SystemdReboot(u32),

}

General command class for ALL commands to the tof C&C instance and readout boards

Each command can carry a 32bit field with further instructionns

Variants

Unknown(u32)

Ping(u32)

Moni(u32)

Power(u32)

SetThresholds(u32)

SetMTConfig(u32)

SetPreampBias(u32)

DataRunStop(u32)

DataRunStart(u32)

StartValidationRun(u32)

GetFullWaveforms(u32)

NoiCalibration(u32)

VoltageCalibration(u32)

TimingCalibration(u32)

DefaultCalibration(u32)

UnspoolEventCache(u32)

SetRBDataBufSize(u32)

TriggerModeForced(u32)

TriggerModeForcedMTB(u32)

SystemdReboot(u32)

Implementations

impl TofCommand

pub fn to_bytearray(&self) -> [u8; 9]

Returns the serialized data stream as byte array

Might be faster thant its sister ::to_bytestream(), however is not a trait, since the return type depends on the size. FIXME - can we somehow make this a trait? It seems we can not return &u8 when we have the corresponding array allocated in the function

pub fn to_bytestream(&self) -> Vec<u8>

pub fn get_value(&self) -> u32

pub fn from_command_code(cc: TofCommandCode, value: u32) -> TofCommand

Generate a TofCommand from the specific bytecode representation

pub fn to_command_code(cmd: &TofCommand) -> Option<TofCommandCode>

Translate a TofCommand into its specific byte representation

pub fn from_tof_packet(packet: &TofPacket) -> Option<TofCommand>