tof_dataclasses::commands
Enum TofCommandCode

#[repr(u8)]
pub enum TofCommandCode {

    Unknown = 0,
    Ping = 1,
    Moni = 2,
    Kill = 4,
    ResetConfigWDefault = 5,
    SubmitConfig = 6,
    SetDataPublisherConfig = 20,
    SetLTBThresholds = 21,
    SetMTConfig = 22,
    SetTofRunConfig = 23,
    SetTofRBConfig = 24,
    SetAnalysisEngineConfig = 27,
    SetPreampBias = 28,
    SetTOFEventBuilderConfig = 29,
    DataRunStop = 30,
    DataRunStart = 31,
    StartValidationRun = 32,
    GetFullWaveforms = 41,
    UnspoolEventCache = 44,
    RBCalibration = 53,
    RestartLiftofRBClients = 60,
    Listen = 70,
    Staging = 71,
    Lock = 80,
    Unlock = 81,
    SendTofEvents = 90,
    NoSendTofEvents = 91,
    SendRBWaveforms = 92,
    NoSendRBWaveforms = 93,
    SetRBChannelMask = 99,
    ShutdownRB = 100,
    ChangeNextRunConfig = 101,
    ShutdownRAT = 102,
    ShutdownRATPair = 103,
    ShutdownCPU = 104,
    UploadConfig = 105,

}

Variants

Unknown = 0

Ping = 1

en empty command just to check if stuff is online

Moni = 2

command code for getting the monitoring data from the component

Kill = 4

Kill myself

ResetConfigWDefault = 5

Reload a default (to be defined) config file

SubmitConfig = 6

Make the current editable config the active config

SetDataPublisherConfig = 20

command code to configure the data publisher thread

SetLTBThresholds = 21

command code for โSet LTB Thresholdsโ

SetMTConfig = 22

command code for โConfigure MTBโ

SetTofRunConfig = 23

command code for chaning general run parameters

SetTofRBConfig = 24

command code for changing RB parameters

SetAnalysisEngineConfig = 27

command code for AnalysisEngineConfig

SetPreampBias = 28

command code for โSet preamp biasโ

SetTOFEventBuilderConfig = 29

Change the settings of the event builder

DataRunStop = 30

command code for โStop Data takingโ

DataRunStart = 31

command code for โStart Data takingโ

StartValidationRun = 32

command code for โStart validation runโ

GetFullWaveforms = 41

command code for โGet all waveformsโ

UnspoolEventCache = 44

command code for โSend the whole event cache over the wireโ

RBCalibration = 53

command code for โRun full calibrationโ

RestartLiftofRBClients = 60

command code for restarting systemd

Listen = 70

command code for putting liftof-cc in listening mode

Staging = 71

command code for putting liftof-cc in staging mode

Lock = 80

lock the cmd dispatcher

Unlock = 81

unlock the cmd dispatcher

SendTofEvents = 90

Enable sending of TOF packets

NoSendTofEvents = 91

Diesable sending of TofEventPacket

SendRBWaveforms = 92

Enable sending of RBWaveform packets

NoSendRBWaveforms = 93

Disable sending of RBWaveform packets

SetRBChannelMask = 99

Enable RB Channel Masks

ShutdownRB = 100

Shutdown RB - send shutdown now to RB

ChangeNextRunConfig = 101

Change the config file for the next run

ShutdownRAT = 102

Shutdown RAT - send shutdown command to 2RBs in the same RAT

ShutdownRATPair = 103

Shutdown a pair of RATs (as always two of them are hooked up to the same PDU channel)

ShutdownCPU = 104

Shutdown the TOF CPU

UploadConfig = 105

Upload a new config file