Enum tof_dataclasses::config::BuildStrategy

pub enum BuildStrategy {
    Unknown,
    Smart,
    Adaptive,
    AdaptiveThorough,
    AdaptiveGreedy,
    WaitForNBoards,
}

Build Strategy

Variants

Unknown

Smart

Adaptive

adjust the number of boards based on nrbes/mtb

AdaptiveThorough

Same as adaptive, but check if the rb events follow the mapping

AdaptiveGreedy

like adaptive, but add usize to the expected number of boards

WaitForNBoards

Implementations

impl BuildStrategy

pub fn to_u8(&self) -> u8