tof_dataclasses::database

Struct DSICard

pub struct DSICard {
    pub dsi_id: i16,
    pub j1_rat_id: Option<i16>,
    pub j2_rat_id: Option<i16>,
    pub j3_rat_id: Option<i16>,
    pub j4_rat_id: Option<i16>,
    pub j5_rat_id: Option<i16>,
}

A DSI card which is plugged into one of five slots on the MTB The DSI card provides the connection to RBs and LTBs and has a subdivision, which is called ‘j’

Fields
dsi_id: i16j1_rat_id: Option<i16>j2_rat_id: Option<i16>j3_rat_id: Option<i16>j4_rat_id: Option<i16>j5_rat_id: Option<i16>
Implementations

impl DSICard

pub fn new() -> Self

pub fn has_rat(&self, r_id: u8) -> bool

True if this RAT box is plugged in to any of the j connectors on this specific DSI card

pub fn get_j(&self, r_id: u8) -> Option<u8>

Get the j connetor for this specific RAT Raises ValueError if the RAT is not connected

pub fn all(conn: &mut SqliteConnection) -> Option<Vec<DSICard>>