Struct tof_dataclasses::events::blob::BlobData

pub struct BlobData {

    pub head: u16,
    pub status: u16,
    pub len: u16,
    pub roi: u16,
    pub dna: u64,
    pub fw_hash: u16,
    pub id: u16,
    pub ch_mask: u16,
    pub event_id: u32,
    pub dtap0: u16,
    pub dtap1: u16,
    pub timestamp: u64,
    pub ch_head: [u16; 9],
    pub ch_adc: [[i16; 1024]; 9],
    pub ch_trail: [u32; 9],
    pub stop_cell: u16,
    pub crc32: u32,
    pub tail: u16,
    pub voltages: [[f64; 1024]; 9],
    pub nanoseconds: [[f64; 1024]; 9],
    pub threshold: [f64; 9],
    pub cfds_fraction: [f64; 9],
    pub ped_begin_bin: [usize; 9],
    pub ped_bin_range: [usize; 9],
    pub pedestal: [f64; 9],
    pub pedestal_sigma: [f64; 9],
    pub peaks: [[usize; 50]; 9],
    pub tdcs: [[f64; 50]; 9],
    pub charge: [[f64; 50]; 9],
    pub width: [[f64; 50]; 9],
    pub height: [[f64; 50]; 9],
    pub num_peaks: [usize; 9],
    pub begin_peak: [[usize; 50]; 9],
    pub end_peak: [[usize; 50]; 9],
    pub spikes: [[usize; 50]; 9],
    pub impedance: f64,

}

Fields
head: u16status: u16len: u16roi: u16dna: u64fw_hash: u16id: u16ch_mask: u16event_id: u32dtap0: u16dtap1: u16timestamp: u64ch_head: [u16; 9]ch_adc: [[i16; 1024]; 9]ch_trail: [u32; 9]stop_cell: u16crc32: u32tail: u16voltages: [[f64; 1024]; 9]nanoseconds: [[f64; 1024]; 9]threshold: [f64; 9]cfds_fraction: [f64; 9]ped_begin_bin: [usize; 9]ped_bin_range: [usize; 9]pedestal: [f64; 9]pedestal_sigma: [f64; 9]peaks: [[usize; 50]; 9]tdcs: [[f64; 50]; 9]charge: [[f64; 50]; 9]width: [[f64; 50]; 9]height: [[f64; 50]; 9]num_peaks: [usize; 9]begin_peak: [[usize; 50]; 9]end_peak: [[usize; 50]; 9]spikes: [[usize; 50]; 9]impedance: f64
Implementations

impl BlobData

pub const SERIALIZED_SIZE: usize = 18_530usize

FIXME - more correctly, this should be “deserialize from readoutboard_blob

pub const HEAD: u16 = 43_690u16

pub const TAIL: u16 = 21_845u16

pub fn new() -> BlobData

pub fn decode_event_id(bytestream: &[u8]) -> u32

! Only decode the event id from a bytestream

The bytestream has to be starting with HEAD

pub fn from_bytestream_experimental(
&mut self,
bytestream: &Vec<u8>,
start_pos: usize,
search_start: bool
) -> usize

! EXPERIMENTAL Initialize the blob from a bytestream.

This is a member here, so this can be done repeatedly without re-allocation of the blob arrays.

THIS VERSION IS DEDICATED TO OPERATE ON THE RB!

Arguments

search_start : automatically look ahead from start_pos unitl HEAD is found

FIXME - this should return Result!

pub fn from_bytestream(
&mut self,
bytestream: &Vec<u8>,
start_pos: usize,
search_start: bool
) -> usize

! Initialize the blob from a bytestream.

This is a member here, so this can be done repeatedly without re-allocation of the blob arrays

Arguments

search_start : automatically look ahead from start_pos unitl HEAD is found

FIXME - this should return Result!

pub fn to_bytestream(&self) -> Vec<u8>

! Serialize the event to a bytestream

pub fn calibrate(&mut self, cal: &[Calibrations; 9])

Apply the calibration for time and voltage.

pub fn remove_spikes(&mut self, spikes: &mut [i32; 10])

pub fn set_threshold(&mut self, thr: f64, ch: usize) -> bool

! Set the threshold and check if the waveform got over threahshold

pub fn set_cfds_fraction(&mut self, fraction: f64, ch: usize)

pub fn set_ped_begin(&mut self, time: f64, ch: usize)

pub fn set_ped_range(&mut self, range: f64, ch: usize)

pub fn subtract_pedestal(&mut self, ch: usize)

pub fn calc_ped_range(&mut self, ch: usize)

pub fn find_cfd_simple(&mut self, peak_num: usize, ch: usize) -> f64

pub fn find_interpolated_time(
&self,
threshold: f64,
idx: usize,
size: usize,
ch: usize
) -> Result<f64, WaveformError>

pub fn find_peaks(&mut self, start_time: f64, window_size: f64, ch: usize)

Find peaks in a given time window (in ns) by comparing the waveform voltages with the given threshold. Minimum peak width is currently hardcoded to be 3 bins in time.

pub fn integrate(
&mut self,
lower_bound: f64,
size: f64,
channel: usize
) -> Result<f64, WaveformError>

pub fn reset(&mut self)

pub fn print(&self)