Enum tof_dataclasses::events::master_trigger::TriggerType

#[repr(u8)]
pub enum TriggerType {

    Unknown = 0,
    Any = 1,
    Track = 2,
    TrackCentral = 3,
    Gaps = 4,
    Gaps633 = 5,
    Gaps422 = 6,
    Gaps211 = 7,
    TrackUmbCentral = 8,
    UmbCube = 21,
    UmbCubeZ = 22,
    UmbCorCube = 23,
    CorCubeSide = 24,
    Umb3Cube = 25,
    Poisson = 100,
    Forced = 101,
    FixedRate = 102,
    ConfigurableTrigger = 200,

}

Variants

Unknown = 0

Any = 1

-> 1-10 โpysicsโ triggers

Track = 2

TrackCentral = 3

Gaps = 4

Gaps633 = 5

Gaps422 = 6

Gaps211 = 7

TrackUmbCentral = 8

UmbCube = 21

-> 20+ โPhilipโs triggersโ Any paddle HIT in UMB + any paddle HIT in CUB

UmbCubeZ = 22

Any paddle HIT in UMB + any paddle HIT in CUB top

UmbCorCube = 23

Any paddle HIT in UMB + any paddle hit in COR + any paddle hit in CUB

CorCubeSide = 24

Any paddle HIT in COR + any paddle HIT in CUB SIDES

Umb3Cube = 25

Any paddle hit in UMB + any three paddles HIT in CUB

Poisson = 100

100 -> Debug triggers

Forced = 101

FixedRate = 102

ConfigurableTrigger = 200

200 -> These triggers can not be set, they are merely the result of what we read out from the trigger mask of the ltb

Implementations

impl TriggerType

pub fn to_u8(&self) -> u8