Struct tof_dataclasses::events::master_trigger::MasterTriggerEvent

pub struct MasterTriggerEvent {
    pub event_status: EventStatus,
    pub event_id: u32,
    pub timestamp: u32,
    pub tiu_timestamp: u32,
    pub tiu_gps32: u32,
    pub tiu_gps16: u16,
    pub crc: u32,
    pub trigger_source: u16,
    pub dsi_j_mask: u32,
    pub channel_mask: Vec<u16>,
    pub mtb_link_mask: u64,
}

An event as observed by the MTB

This is condensed to the most crucial information

FIXME : implementation of absolute time

Fields
event_status: EventStatusevent_id: u32timestamp: u32

Internal timestamp at the time of trigger (1 unit = 10 ns) Free running counter, rolling over every ~42 seconds

tiu_timestamp: u32

Timestamp at the edge of the TIU GPS (1 unit = 10 ns)

tiu_gps32: u32

Second received from the TIU (format?)

tiu_gps16: u16crc: u32trigger_source: u16

Trigger source:

dsi_j_mask: u32channel_mask: Vec<u16>mtb_link_mask: u64
Implementations

impl MasterTriggerEvent

pub const VERSION: u8 = 3u8

Implementation version, might roughly correspond to fw version

pub fn new() -> Self

pub fn get_rb_link_ids(&self) -> Vec<u8>

Get the RB link IDs according to the mask

pub fn get_trigger_hits(&self) -> Vec<(u8, u8, (u8, u8), LTBThreshold)>

Get the combination of triggered DSI/J/CH on the MTB which formed the trigger. This does not include further hits which fall into the integration window. For those, se rb_link_mask

The returned values follow the TOF convention to start with 1, so that we can use them to look up LTB ids in the db.

Returns

Vec<(hit)> where hit is (DSI, J, (CH,CH), threshold)

pub fn get_timestamp_gps48(&self) -> u64

combine the tiu gps 16 and 32bit timestamps into a 48bit timestamp

pub fn get_timestamp_abs48(&self) -> u64

Get absolute timestamp as sent by the GPS

pub fn get_trigger_sources(&self) -> Vec<TriggerType>

Get the trigger sources from trigger source byte In case of the custom (configurable triggers, this

will only return “ConfigurableTrigger” since the MTB does not know about these triggers as individual types

pub fn get_global_trigger_soures(&self) -> Vec<TriggerType>

Returns the trigger types which have to be defined as “global”

Global triggers will force a readout of all panels and can be operated in conjuction with the set trigger