Module tof_dataclasses::events::rb_event

Readoutboard binary event formats, former denoted as BLOB (binary large object)

The structure is the following

RBEventHeader - timestamp, status, len of event, but no channel data. This represents compression level 2

RBEvent - still “raw” event, however, with modified fields (removed superflous ones, changed meaning of some others) Each RBEvent has a header and a body which is the channel data. Data in this form represents compression level 1

RBWaveform - a single waveform from a single RB. This can be used to deconstruct TofEvents so that the flight computer does not struggle with the large packet size.

features: “random” - provides “::from_random” for all structs allowing to populate them with random data for tests.

Structs

RBEvent

Event data for each individual ReadoutBoard (RB)

RBEventHeader

The RBEvent header gets generated once per event per RB. Contains information about event id, timestamps, etc.

RBMissingHitDeprecated

Debug information for missing hits.

RBWaveform

Functions

unpack_traces_f32

unpack_traces_f64

Get the traces for a set of RBEvents