Struct tof_dataclasses::events::rb_event::RBEventHeader

pub struct RBEventHeader {
    pub rb_id: u8,
    pub event_id: u32,
    pub status_byte: u8,
    pub channel_mask: u16,
    pub stop_cell: u16,
    pub ch9_amp: u16,
    pub ch9_freq: u16,
    pub ch9_phase: u32,
    pub fpga_temp: u16,
    pub timestamp32: u32,
    pub timestamp16: u16,
}

The RBEvent header gets generated once per event per RB. Contains information about event id, timestamps, etc.

Fields
rb_id: u8event_id: u32status_byte: u8channel_mask: u16stop_cell: u16ch9_amp: u16ch9_freq: u16ch9_phase: u32fpga_temp: u16timestamp32: u32timestamp16: u16
Implementations

impl RBEventHeader

pub fn new() -> Self

pub fn set_sine_fit(&mut self, input: (f32, f32, f32))

pub fn get_sine_fit(&self) -> (f32, f32, f32)

pub fn extract_eventid_from_rbheader(stream: &Vec<u8>) -> u32

Only get the eventid from a binary stream

pub fn is_event_fragment(&self) -> bool

pub fn drs_lost_trigger(&self) -> bool

pub fn lost_lock(&self) -> bool

pub fn lost_lock_last_sec(&self) -> bool

pub fn is_locked(&self) -> bool

pub fn is_locked_last_sec(&self) -> bool

pub fn parse_status(&mut self, status_bytes: u16)

extract lock, drs busy and fpga temp from status field

pub fn get_fpga_temp(&self) -> f32

Get the temperature value (Celsius) from the fpga_temp adc.

pub fn has_ch9(&self) -> bool

Get the entire header from a full binary representation of the raw RBEventMemoryView encoded in a binary stream

pub fn get_channels(&self) -> Vec<u8>

Decode the channel mask into channel ids.

The channel ids inside the memory representation of the RB Event data (“blob”) are from 0-7

We keep ch9 seperate.

pub fn get_nchan(&self) -> usize

Get the number of data channels + 1 for ch9

pub fn get_timestamp48(&self) -> u64