Struct tof_dataclasses::events::rb_eventmemoryview::RBEventMemoryView

pub struct RBEventMemoryView {

    pub head: u16,
    pub status: u16,
    pub len: u16,
    pub roi: u16,
    pub dna: u64,
    pub fw_hash: u16,
    pub id: u16,
    pub ch_mask: u16,
    pub event_id: u32,
    pub dtap0: u16,
    pub dtap1: u16,
    pub timestamp_32: u32,
    pub timestamp_16: u16,
    pub ch_head: [u16; 9],
    pub ch_adc: [[u16; 1024]; 9],
    pub ch_trail: [u32; 9],
    pub stop_cell: u16,
    pub crc32: u32,
    pub tail: u16,

}

๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!

RBEventMemoryView is the closest representation of actual RB binary data in memory, with a fixed number of channels at compile time, optimized for speed by using fixed (at compile time) sizes for channels and sample size

FIXME - the channel mask is only one byte, and we can get rid of 3 bytes for the DNA

Fields
head: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
status: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
len: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
roi: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
dna: u64
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
fw_hash: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
id: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
ch_mask: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
event_id: u32
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
dtap0: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
dtap1: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
timestamp_32: u32
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
timestamp_16: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
ch_head: [u16; 9]
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
ch_adc: [[u16; 1024]; 9]
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
ch_trail: [u32; 9]
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
stop_cell: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
crc32: u32
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!
tail: u16
๐Deprecated since 0.7.2: RBEvent is sufficient to fulfill all our needs!

Implementations

impl RBEventMemoryView

pub fn new() -> Self

pub fn decode_event_id(bytestream: &[u8]) -> Result<u32, SerializationError>

pub fn get_active_data_channels(&self) -> Vec<u8>

pub fn get_n_datachan(&self) -> u8