Struct tof_dataclasses::events::tof_hit::TofHit

pub struct TofHit {

    pub paddle_id: u8,
    pub time_a: f16,
    pub time_b: f16,
    pub peak_a: f16,
    pub peak_b: f16,
    pub charge_a: f16,
    pub charge_b: f16,
    pub timestamp32: u32,
    pub timestamp16: u16,
    pub ctr_etx: u8,
    pub charge_min_i: u16,
    pub pos_across: u16,
    pub t0: u16,
    pub reserved: u8,
    pub version: ProtocolVersion,
    pub baseline_a: f16,
    pub baseline_a_rms: f16,
    pub baseline_b: f16,
    pub baseline_b_rms: f16,
    pub phase: f16,
    pub valid: bool,
    pub ftime_a: f32,
    pub ftime_b: f32,
    pub fpeak_a: f32,
    pub fpeak_b: f32,
    pub paddle_len: f32,

}

Comprehensive paddle information

Results of the (online) waveform analysis

A and B are the different ends of the paddle

Fields
paddle_id: u8

The ID of the paddle in TOF notation (1-160)

time_a: f16time_b: f16peak_a: f16peak_b: f16charge_a: f16charge_b: f16timestamp32: u32timestamp16: u16ctr_etx: u8charge_min_i: u16pos_across: u16

Reconstructed particle interaction position across the paddle

t0: u16

Reconstructed particle interaction time

reserved: u8version: ProtocolVersionbaseline_a: f16baseline_a_rms: f16baseline_b: f16baseline_b_rms: f16phase: f16valid: boolftime_a: f32ftime_b: f32fpeak_a: f32fpeak_b: f32paddle_len: f32
Implementations

impl TofHit

pub fn new() -> Self

pub fn get_pid(paddle_end_id: u16) -> u8

Get the (official) paddle id

Convert the paddle end id following the convention

A-side : paddle id + 1000 B-side : paddle id + 2000

FIXME - maybe return Result?

pub fn add_peak(&mut self, peak: &Peak)

pub fn get_pos(&self) -> f32

Calculate the position across the paddle from the two times at the paddle ends

This will be measured from the A side

pub fn get_t0(&self) -> f32

Calculate the interaction time based on the peak timings measured at the paddle ends A and B

That this works, the lenght of the paddle has to be set before.

pub fn get_edep(&self) -> f32

pub fn get_time_a(&self) -> f32

pub fn set_time_a(&mut self, t: f32)

pub fn get_time_b(&self) -> f32

pub fn set_time_b(&mut self, t: f32)

pub fn get_peak_a(&self) -> f32

pub fn set_peak_a(&mut self, p: f32)

pub fn get_peak_b(&self) -> f32

pub fn set_peak_b(&mut self, p: f32)

pub fn get_charge_a(&self) -> f32

pub fn set_charge_a(&mut self, c: f32)

pub fn get_charge_b(&self) -> f32

pub fn set_charge_b(&mut self, c: f32)

pub fn get_bl_a(&self) -> f32

pub fn get_bl_b(&self) -> f32

pub fn get_bl_a_rms(&self) -> f32

pub fn get_bl_b_rms(&self) -> f32