tof_dataclasses

Module io

Dataio - readino/writing of different types

of TOF data products.

TofPacketReader/Writer - sequentially read/write TofPackets from/to a file on disk. -> upcoming: Will connect to a network socket

RobinReader: Read (old) RB data files, where the file is simply a dump of the internal buffers (“RBEventMemoryView”).

RBEventMemoryStreamer: Walk over “raw” RBEvents representations (“RBEventMemoryView”) and extract RBEvents

Structs

RBEventMemoryStreamer

Emit RBEvents from a stream of bytes from RBMemory

RobinReader

Read RB binary (robin) files. These are also known as “blob” files

TofPacketReader

Read serialized TofPackets from an existing file or directory

TofPacketWriter

Write TofPackets to disk.

Enums

FileType

Types of files

Traits

PacketReader

Generics for packet reading (TofPacket, Telemetry packet,…) FIXME - not implemented yet

Functions

get_califilename

A standardized name for calibration files saved by the liftof suite

get_runfilename

A standardized name for regular run files saved by the liftof suite

get_utc_date

Create date string in YYMMDD format

get_utc_timestamp

Get a human readable timestamp

read_file

Read an entire file into memory

summarize_toffile

Take a .tof.gaps file with TofEvents and keep all packets, but reduce the TofEvents to TofEventSuammry to conserve space.