Struct tof_dataclasses::io::RBEventMemoryStreamer

pub struct RBEventMemoryStreamer {

    pub stream: Vec<u8>,
    pub check_channel_errors: bool,
    pub tp_sender: Option<Sender<TofPacket>>,
    pub is_depleted: bool,
    pub calc_crc32: bool,
    pub request_mode: bool,
    pub request_cache: VecDeque<(u32, u8)>,
    pub event_map: HashMap<u32, (usize, usize)>,
    pub first_evid: u32,
    pub last_evid: u32,
    pub last_event_complete: bool,
    pub last_event_pos: usize,
    pub is_behind_by: usize,
    pub is_ahead_by: usize,
    /* private fields */

}

Emit RBEvents from a stream of bytes from RBMemory

The layout of the stream has to have the fpga fw memory layout.

This provides a next() method to act as a generator for RBEvents

Fields
stream: Vec<u8>

Raw stream read out from the RB buffers.

check_channel_errors: bool

Error checking mode - check error bits for channels/cells

tp_sender: Option<Sender<TofPacket>>

An optional crossbeam::channel Sender, which will allow to send TofPackets

is_depleted: boolcalc_crc32: bool

Calculate the crc32 checksum for the channels everytime next() is called

request_mode: boolrequest_cache: VecDeque<(u32, u8)>event_map: HashMap<u32, (usize, usize)>

an index for the events in the stream this links eventid and start position in the stream together

first_evid: u32last_evid: u32last_event_complete: boollast_event_pos: usizeis_behind_by: usize

When in request mode, number of events the last event in the stream is behind the first request

is_ahead_by: usize

When in request mode, number of events the last event in the stream is ahead the last request

Implementations

impl RBEventMemoryStreamer

pub fn new() -> Self

pub fn create_event_index(&mut self)

Create the event index, which is a map of event ids and position

length in the stream

pub fn print_event_map(&self)

pub fn init_sender(&mut self, tp_sender: Sender<TofPacket>)

pub fn send_all(&mut self)

pub fn add(&mut self, stream: &Vec<u8>, nbytes: usize)

pub fn consume(&mut self, stream: &mut Vec<u8>)

Take in a stream by consuming it, that means moving This will avoid clones.

pub fn seek_next_header(&mut self, header: u16) -> bool

Headers are expected to be a 2byte signature, e.g. 0xaaaa.

Arguments:

header : 2byte header.

Returns

success : header found

pub fn next_tofpacket(&mut self) -> Option<TofPacket>

pub fn get_event_at_pos_unchecked( &mut self, replace_channel_mask: Option<u16> ) -> Option<RBEvent>

Retrive an RBEvent from a certain position

pub fn get_event_at_id( &mut self, event_id: u32, replace_channel_mask: Option<u16> ) -> Option<RBEvent>