tof_dataclasses::io

Struct TofPacketWriter

pub struct TofPacketWriter {
    pub file: File,
    pub file_path: String,
    pub pkts_per_file: usize,
    pub mbytes_per_file: usize,
    pub file_type: FileType,
    pub file_name: String,
    /* private fields */
}

Write TofPackets to disk.

Operates sequentially, packets can be added one at a time, then will be synced to disk.

Fields
file: Filefile_path: String

location to store the file

pkts_per_file: usize

The maximum number of packets for a single file. Ater this number is reached, a new file is started.

mbytes_per_file: usize

The maximum number of (Mega)bytes per file. After this a new file is started

file_type: FileType

add timestamps to filenames

file_name: String
Implementations

impl TofPacketWriter

pub fn new(file_path: String, file_type: FileType) -> Self

Instantiate a new PacketWriter

Arguments

file_prefix : Prefix file with this string. A continuous number will get appended to control the file size.

file_type : control the behaviour of how the filename is assigned.

pub fn get_file(&self) -> File

pub fn add_tof_packet(&mut self, packet: &TofPacket)

Induce serialization to disk for a TofPacket