Struct tof_dataclasses::ipbus::IPBus

pub struct IPBus {
    pub socket: UdpSocket,
    pub target_address: String,
    pub packet_type: IPBusPacketType,
    pub pid: u16,
    pub expected_pid: u16,
    pub last_pid: u16,
    pub buffer: [u8; 128],
}

Implementation of an IPBus control packet

Fields
socket: UdpSockettarget_address: Stringpacket_type: IPBusPacketTypepid: u16

IPBus Packet ID - this is then NEXT pid which will be sent

expected_pid: u16last_pid: u16buffer: [u8; 128]
Implementations

impl IPBus

pub fn new(target_address: String) -> Result<Self>

pub fn connect(target_address: &String) -> Result<UdpSocket>

Connect to MTB Utp socket

This will try a number of options to bind to the local port.

Arguments

target_address : IP/port of the target probably some kind of FPGA

pub fn reconnect(&mut self) -> Result<()>

Reconnect to the same address after timeout

pub fn receive(&mut self) -> Result<usize>

Receive number_of_bytes from UdpSocket and sleep after to avoid too many queries

pub fn send(&mut self, data: &Vec<u8>) -> Result<()>

Receive number_of_bytes from UdpSocket and sleep after to avoid too many queries

pub fn get_status(&mut self) -> Result<(), Box<dyn Error>>

Send a ipbus status packet and receive the response

Inspect self.buffer after the call if interested in the result.

pub fn get_pid_from_current_buffer(&self) -> u16

pub fn realign_packet_id(&mut self) -> Result<(), Box<dyn Error>>

Set the packet id to that what is expected from the targetr

pub fn buffer_is_status(&self) -> bool

pub fn get_target_next_expected_packet_id( &mut self ) -> Result<u16, Box<dyn Error>>

Get the packet id which is expected by the target

pub fn read_multiple( &mut self, addr: u32, nwords: usize, increment_addr: bool ) -> Result<Vec<u32>, Box<dyn Error>>

Multiple read operations with a single UDP request

Read either the same register multiple times or read from incrementing register addresses

Arguments:

addr : register addresss to read from

nwords : number of read operations

increment_addr : if true, increment the register address after each read operation

pub fn read(&mut self, addr: u32) -> Result<u32, Box<dyn Error>>

Read a single value from a register

Arguments:

addr : register address to be read from

pub fn write(&mut self, addr: u32, data: u32) -> Result<(), Box<dyn Error>>

Write a single value to a register

Arguments

addr : target register address

data : word to write in register