Module tof_dataclasses::manifest

Entities of the tof, that is LocalTriggerBoard, ReadoutBoard, PB, MTB…

This provides structs to hold their metainformation, methods to save/load them from the gaps database as well as to serialize them locally.

Structs

DSICard

LocalTriggerBoard

MTBChannel

Summary of DSI/J/LTBCH (0-319)

Paddle

A represnetation of a TOF paddle

PaddleEnd

Panel

This represents an entire TOF panel (an assembly of paddles)

RAT

ReadoutBoard

Enums

PaddleEndIdentifier

PaddleEndLocation

Functions

get_all_mtbchannels

get_dsi_from_sqlite

get_dsi_j_ltbch_vs_rbch_map

Dsi -> J -> (RBID,RBCH)

get_hg_lg_map

get_lg_hg_map

get_linkid_rbid_map

Create a mapping of mtb link ids to rb ids

get_ltbs_from_sqlite

get_pid_from_pend

get_rbs_from_sqlite