Module tof_dataclasses::monitoring

Tof housekeeping/monitoring

Contains structs to hold monitoring information for the different parts of the TOF, e.g. RB,LTB,MTB

An overview of the sensors in the GAPS TOF can be found in the GAPS wiki

Re-exports

pub use crate::series::MoniSeries;

pub use crate::series::PAMoniDataSeries;

pub use crate::series::PBMoniDataSeries;

pub use crate::series::LTBMoniDataSeries;

pub use crate::series::RBMoniDataSeries;

Structs

CPUMoniData

LTBMoniData

Sensors on the LTB

MtbMoniData

Monitoring the MTB

PAMoniData

Preamp temperature and bias data

PBMoniData

Sensors on the power boards (PB)

RBMoniData

Sensors on the individual RB

Traits

MoniData

Monitoring data shall share the same kind of interface.