tof_dataclasses::monitoring
Trait MoniData

pub trait MoniData {
    // Required methods
    fn get_board_id(&self) -> u8;
    fn get(&self, varname: &str) -> Option<f32>;
    fn keys() -> Vec<&'static str>;
}

Monitoring data shall share the same kind of interface.

Required Methods

fn get_board_id(&self) -> u8

Monitoring data is always tied to a specific board. This might not be its own board, but maybe the RB the data was gathered from This is an unique identifier for the monitoring data

fn get(&self, varname: &str) -> Option<f32>

Access the (data) members by name

fn keys() -> Vec<&'static str>

A list of the variables in this MoniData

Object Safety

This trait is not object safe.