tof_dataclasses::packets::packet_type
Enum PacketType

#[repr(u8)]
pub enum PacketType {

    Unknown = 0,
    RBEvent = 20,
    TofEvent = 21,
    RBWaveform = 22,
    TofEventSummary = 23,
    HeartBeatDataSink = 40,
    MasterTrigger = 60,
    TriggerConfig = 61,
    MTBHeartbeat = 62,
    EVTBLDRHeartbeat = 63,
    RBChannelMaskConfig = 64,
    TofRBConfig = 68,
    AnalysisEngineConfig = 69,
    RBEventHeader = 70,
    TOFEventBuilderConfig = 71,
    DataPublisherConfig = 72,
    TofRunConfig = 73,
    CPUMoniData = 80,
    MonitorMtb = 90,
    RBMoniData = 100,
    PBMoniData = 101,
    LTBMoniData = 102,
    PAMoniData = 103,
    RBEventMemoryView = 120,
    RBCalibration = 130,
    TofCommand = 140,
    TofCommandV2 = 141,
    TofResponse = 142,
    RBCommand = 150,
    RBPing = 160,
    PreampBiasConfig = 161,
    RunConfig = 162,
    LTBThresholdConfig = 163,
    TofDetectorStatus = 171,
    ConfigBinary = 201,
    LiftofRBBinary = 202,
    LiftofBinaryService = 203,
    LiftofCCBinary = 204,
    RBCalibrationFlightV = 210,
    RBCalibrationFlightT = 211,
    BfswAckPacket = 212,
    MultiPacket = 255,

}

Types of serializable data structures used throughout the tof system

Variants

Unknown = 0

RBEvent = 20

TofEvent = 21

RBWaveform = 22

TofEventSummary = 23

HeartBeatDataSink = 40

MasterTrigger = 60

TriggerConfig = 61

MTBHeartbeat = 62

EVTBLDRHeartbeat = 63

RBChannelMaskConfig = 64

TofRBConfig = 68

AnalysisEngineConfig = 69

RBEventHeader = 70

TOFEventBuilderConfig = 71

DataPublisherConfig = 72

TofRunConfig = 73

CPUMoniData = 80

MonitorMtb = 90

RBMoniData = 100

PBMoniData = 101

LTBMoniData = 102

PAMoniData = 103

RBEventMemoryView = 120

RBCalibration = 130

TofCommand = 140

TofCommandV2 = 141

TofResponse = 142

RBCommand = 150

RBPing = 160

PreampBiasConfig = 161

RunConfig = 162

LTBThresholdConfig = 163

TofDetectorStatus = 171

ConfigBinary = 201

LiftofRBBinary = 202

LiftofBinaryService = 203

LiftofCCBinary = 204

RBCalibrationFlightV = 210

RBCalibrationFlightT = 211

BfswAckPacket = 212

A klude which allows us to send bfsw ack packets through the TOF system

MultiPacket = 255

a MultiPacket consists of other TofPackets