Struct tof_dataclasses::packets::TofPacket

pub struct TofPacket {
    pub packet_type: PacketType,
    pub payload: Vec<u8>,
    pub is_multi_packet: bool,
    pub no_write_to_disk: bool,
    pub no_send_over_nw: bool,
    pub creation_time: Instant,
    pub valid: bool,
}

The most basic of all packets

A type and a payload. This wraps all other packets.

Format when in bytestream HEAD : u16 PAYLOAD_SIZE : u32 PYALOAD : [u8;PAYLOAD_SIZE] TAIL : u16

=> Fixed size is 13

Fields
packet_type: PacketTypepayload: Vec<u8>is_multi_packet: bool

packets which contain multiple of the same packets

no_write_to_disk: bool

mark a packet as not eligible to be written to disk

no_send_over_nw: bool

mark a packet as not eligible to be sent over network FIXME - future extension

creation_time: Instant

creation_time for the instance

valid: bool
Implementations

impl TofPacket

pub const PRELUDE_SIZE: usize = 7usize

pub fn new() -> Self

pub fn unpack_rbevent(&self) -> Result<RBEvent, Box<dyn Error>>

Unpack possible content

pub fn has_timed_out(&self) -> bool

Event can time out after specified time

pub fn age(&self) -> u64

pub fn get_n_packets(&self) -> u32