Struct tof_dataclasses::run::RunConfig

pub struct RunConfig {
    pub runid: u32,
    pub is_active: bool,
    pub nevents: u32,
    pub nseconds: u32,
    pub tof_op_mode: TofOperationMode,
    pub trigger_poisson_rate: u32,
    pub trigger_fixed_rate: u32,
    pub data_type: DataType,
    pub rb_buff_size: u16,
}

Readoutboard configuration for a specific run

Fields
runid: u32

an unique identifier for this run

is_active: bool

start/stop run

This might get deprecated in a future version!

nevents: u32

limit run to number of events

nseconds: u32

limit run time to number of seconds

tof_op_mode: TofOperationMode

tof operation mode - either โStreamAnyโ, โRequestReplyโ or โRBHighThroughputโ

trigger_poisson_rate: u32

if different from 0, activate RB self trigger in poisson mode

trigger_fixed_rate: u32

if different from 0, activate RB self trigger with fixed rate setting

data_type: DataType

Either โPhysicsโ or a calibration related data type, e.g. โVoltageCalibrationโ.

This might get deprecated in a future version!

rb_buff_size: u16

The value when the readout of the RB buffers is triggered. This number is in size of full events, which correspond to 18530 bytes. Maximum buffer size is a bit more than 3000 events. Smaller buffer allows for a more snappy reaction, but might require more CPU resources (on the board)

Implementations

impl RunConfig

pub const VERSION: &'static str = "1.5"

pub fn new() -> Self

pub fn set_forever(&mut self)

๐Deprecated since 0.8.3: Public argument does not need setter. Semantics need to be clear though.

pub fn runs_forever(&self) -> bool

๐Deprecated since 0.8.3: Public argument does not need setter. Semantics need to be clear though.