tof_dataclasses
Module series

Compact data structures for analysis. Won’t get serialized. Summarize data of all boards and provide a more consice table-like structure for analysis. Currently under a lot of development. The goal is to get something which easily translates to a polars data frame.

Structs

CPUMoniDataSeries

LTBMoniDataSeries

MtbMoniDataSeries

PAMoniDataSeries

PBMoniDataSeries

RBMoniDataSeries

Traits

MoniSeries

A MoniSeries is a collection of (primarily) monitoring data, which comes from multiple senders. E.g. a MoniSeries could hold RBMoniData from all 40 ReadoutBoards.