Crate tof_dataclasses

Tof dataclasses

This crate provides tof related dataclasses for

events

network i/o wrappers (“packets”) for classes

function related constants

calibration

commands/responses

TODO: alerts

features:

random - allow random number generated data classes for testing

database - access the SQLite data base for advanced paddle mapping

Modules

analysis

Implementations of analysis engine This is based on the original code by J.Zweerink

calibrations

The ReadoutBoardCalibration is the first stage in a 3 stage calibration process for the tof

commands

Commmands which can be issued to the various components of the tof system.

constants

Global constants for tof operations

errors

Specific error types

events

Events

io

Dataio - readino/writing of different types of TOF data products.

manifest

Entities of the tof, that is LocalTriggerBoard, ReadoutBoard, PB, MTB…

monitoring

Tof housekeeping/monitoring

packets

Packets are a way to send data over the network.

run

Provides run configuration and settings for RBs and Tof CPU

serialization

Serialization/Deserialization helpers

threading

Traits

FromRandom

Create structures filled with random number to be used for testing and benchmarking

Type Aliases

DsiLtbRBMapping

A type for the master trigger mappings

RBChannelPaddleEndIDMap

A type for the mappings of RB channels - paddle edn ids Paddle end ids are the paddle id + 1000 for A and