Crate tof_dataclasses

Tof dataclasses

This crate provides tof related dataclasses for

events

network i/o wrappers (“packets”) for classes

function related constants

calibration

commands/responses

TODO: alerts

features:

random - allow random number generated data classes for testing

database - access a data base for advanced paddle mapping, readoutboard and ltb information etc.

caraspace - register TofPacket through the caraspace library This allows to write TofPackets to frames, which will ultimatly allow them to write them to caraspace files

Re-exports

pub use version::ProtocolVersion;

Modules

alerts

The alerts are just clues to trigger an individual to respond to a certain situation

analysis

Implementations of analysis engine This is based on the original code by J.Zweerink

calibrations

The ReadoutBoardCalibration is the first stage in a 3 stage calibration process for the tof

caraspace

Register TofPacket with the caraspace system

commands

Commmands which can be issued to the various components of the tof system.

constants

Global constants for TOF operations

database

Database access & entities of the TOF

errors

Specific error types

events

Events

heartbeats

Heartbeats - regularly (or on demand) sent software monitoring data

io

Dataio - readino/writing of different types

ipbus

Implementation of the IPBus protocoll for GAPS

monitoring

Tof housekeeping/monitoring

packets

Packets are a way to send data over the network.

serialization

Serialization/Deserialization helpers

series

Compact data structures for analysis. Won’t get serialized. Summarize data of all boards and provide a more consice table-like structure for analysis. Currently under a lot of development. The goal is to get something which easily translates to a polars data frame.

status

Detector status indicators

threading

Thread control structures FIXME - this should go to liftof-lib

version

Type Aliases

DsiLtbRBMapping

A type for the master trigger mappings Dsi -> J -> (RBID,RBCH)

RBChannelPaddleEndIDMap

A type for the mappings of RB channels - paddle edn ids Paddle end ids are the paddle id + 1000 for A and

RbChPidMapping