Module tof_dataclasses::analysis

Implementations of analysis engine This is based on the original code by J.Zweerink

Functions

calc_edep_simple

An approximation to calculate the energy deposition as used by Philip/Jamie/Jeff

calculate_pedestal

The pedestal is the baseline of the waveform

cfd_simple

Find the onset time of a peak with a constant fraction discrimination method.

find_peaks

Find peaks in a given time window (in ns) by comparing the waveform voltages with the given threshold. Minimum peak width is currently hardcoded to be 3 bins in time.

find_peaks_zscore

Z-scores peak finding algorithm

get_max_bin

Return the bin with the maximum ADC value

get_paddle_t0

Calculate the interaction time based on the peak timings measured at the paddle ends A and B

integrate

Integrate a waveform

interpolate_time

pos_across

Calculate the distance from the A side We will Always use the A side to measure “pos_accross”

time2bin

Given a time in ns, find the bin most closely corresponding to that time