tof_dataclasses
Module commands

Commmands which can be issued to the various components of the tof system.

Re-exports

pub use factory::*;

Modules

config

Payloads for commands that configure an entity of the TOF system.

factory

The actual command list - one function per ocmmand

Structs

TofCommandV2

A general command class with an arbitrary payload

Enums

TofCommand

General command class for ALL commands to the tof C&C instance and readout boards

TofCommandCode

TofOperationMode

How to operate the readout Default mode is to request events from the MasterTrigger. However, we can also stream all the waveforms. CAVEAT: For the whole tof, this will cap the rate at 112 Hz, because of the capacity of the switches.

TofResponse

Each TofCommand triggers a TofResponse in reply

TofResponseCode

TofReturnCode