Module tof_dataclasses::commands

Commmands which can be issued to the various components of the tof system.

Here is a comprehensive list (Sydney)

Power on/off to PBs+RBs+LTBs+preamps (all at once) or MT

Power on/off to LTB or preamp < 2/day Command to power on/off various components (to TOF -> to RB) 5 B:

RBsetup ? Command to run rbsetup on a particular RB (to TOF -> to RBs) 8 B:

Set Thresholds < 3/day Command to set a threshold level on all LTBs (to TOF -> to RBs) 8 B:

Set MT Config 1/run, <10/day? Command to set MT trigger config (to TOF -> to MT) 4 B:

Start Validation Run 1/run, <10/day? Command to take a small amount of data (some number E events, I

360xE full waveforms (from TOF)

Start Data-Taking Run 1/run, <10/day? Command to take regular data (to TOF -> to RBs)

Reduced data packet (from Flight computer)

Stop Run < 1/run, < 10/day Command to stop a run (to TOF -> to RBs) 2 B = command name 6

Voltage Calibration Runs 1/day Command to take 2 voltage calibration runs (to TOF -> to RBs) 12 B:

Timing Calibration Run 1/day Command to take a timing calibration run (to TOF -> to RBs) 8 B:

Create New Calibration File 1/day Command to create a new calibration file using data from the three

Each command will be answered by a specific response. The responses consists of a class, TofResponse together with a 32bit response code.

Structs

TofCommandV2

A general command class with an arbitrary payload

Enums

TofCommand

General command class for ALL commands to the tof C&C instance and readout boards

TofCommandCode

TofOperationMode

How to operate the readout Default mode is to request events from the MasterTrigger. However, we can also stream all the waveforms. CAVEAT: For the whole tof, this will cap the rate at 112 Hz, because of the capacity of the switches.

TofResponse

Each TofCommand triggers a TofResponse in reply

TofResponseCode