tof_dataclasses::commands

Module factory

The actual command list - one function per ocmmand

These are factory functions and each of them will return a single command

Functions

change_datapublisherconfig

Change the data publisher config part of the config file

change_tofeventbuilderconfig

Change the EventBuilderSettings in the config file

change_tofrbconfig

change_tofrunconfig

Change the data publisher config part of the config file

change_triggerconfig

Change the MTBSettings in the config file with relevant trigger settings

get_ratpdumap_hardcoded

A hardwired map of PDU #id PDUCHANNEL #id to (RAT,RAT)

get_ratrbmap_hardcoded

A hardwired map of RAT -> (RB1, RB2)

get_rbratmap_hardcoded

A hardwired map of RB -> RAT

rb_calibration

Run a calibration of all RBs

restart_liftofrb

Restart the liftof-rb clients on the given boards

shutdown_all_rbs

Send the ‘sudo shutdown now’ command to ALL RBs

shutdown_rat

Send the ‘sudo shutdown now’ command to all RBs in a RAT

shutdown_ratpair

Send the ‘sudo shutdown now’ command to all RBs in a RAT

shutdown_rb

Send the ‘sudo shutdown now’ command to a single RB

shutdown_tofcpu

Send the ’sudo shutdown now command to the TOF main computer (“TOFCPU”)

start_run

Trigger the start of a new data run with the next active config

stop_run

Stop the current active run and idle