Module tof_dataclasses::constants

Global constants for tof operations

Constants

EVENT_BUILDER_EVID_CACHE_SIZE

Limit the size of the evids the event builder is currently waiting to get paddles for.

EVENT_CACHE_SIZE

Limit the size of the internanl event cache

EVENT_TIMEOUT

How long to wait for paddles packets for each event in seconds. This does highly depend on the frequency with which the readoutboards are emitting For now, lets use 30s.

EXP_N_PADDLES_PER_EVENT

Average number of paddle packets per event

MASK_CMD_8BIT

Masks for 32 bits commands (byte packets)

MASK_CMD_16BIT

MASK_CMD_24BIT

MASK_CMD_32BIT

MAX_NBOARDSDeprecated

The maximum number of supported readout boards

MAX_NUM_PEAKS

The maximum number of detectable peaks in a waveform

MAX_TRIGGER_RATE

Expected maximum trigger rate

NCHN

Number of Channels on the readoutboards

NWORDS

Number of entries for each waveform (voltage and timing each)

PADDLE_PACKET_CACHE_SIZE

Limit the size of the internal paddle packet cache

PAD_CMD_32BIT

Padding for 32 bits commands (byte packets)

TOFPACKET_TIMEOUT

When should TofPackets time out? (sec)