Module tof_dataclasses::errors

Specific error types

Enums

AnalysisError

Errors when converting events to PaddlePackets or more general, when doing any kind of analysis

CalibrationError

DecodingError

EventError

IPBusError

IPBus provides a package format for sending UDP packets with a header. This is used by the MTB to send its packets over UDP

MasterTriggerError

Error to be used for issues with the communication to the MTB.

PacketError

SensorError

SerializationError

This error shall be thrown whenver there is an issue in the de(serialization), e.g. the from_bytestream methods.

UserError

WaveformError

Problems in waveform analysis