tof_dataclasses::events::rb_event

Function unpack_traces_f64

pub fn unpack_traces_f64(events: &Vec<RBEvent>) -> Vec<Vec<Vec<f64>>>

Get the traces for a set of RBEvents

This will return a cube of The sice of this cube will be fixed in two dimensions, but not the third

The rationale of this is to be able to quickly calculate means over all channels.

Shape [ch:9][nevents][adc_bin:1024]

Args:

events - events to get the traces from