tof_dataclasses::serialization
Trait Serialization

pub trait Serialization {
    const HEAD: u16;
    const TAIL: u16;
    const SIZE: usize = 0usize;

    // Required method
    fn from_bytestream(
        bytestream: &Vec<u8>,
        pos: &mut usize,
    ) -> Result<Self, SerializationError>
       where Self: Sized;

    // Provided methods
    fn verify_fixed(
        stream: &Vec<u8>,
        pos: &mut usize,
    ) -> Result<(), SerializationError> { ... }
    fn from_tofpacket(packet: &TofPacket) -> Result<Self, SerializationError>
       where Self: Sized { ... }
    fn to_bytestream(&self) -> Vec<u8> { ... }
    fn from_slice(
        _slice: &[u8],
        _start_pos: usize,
    ) -> Result<Self, SerializationError>
       where Self: Sized { ... }
    fn to_slice(&self) -> &[u8] 
       where Self: Sized { ... }
}

Encode/decode structs to Vec::<u8> to write to a file or send over the network

Required Associated Constants

const HEAD: u16

const TAIL: u16

Provided Associated Constants

const SIZE: usize = 0usize

The SIZE is the size of the serialized bytestream INCLUDING 4 bytes for head and tail bytes. In case the struct does NOT HAVE a fixed size, SIZE will be 0 (so default value of the trait

Required Methods

fn from_bytestream( bytestream: &Vec<u8>, pos: &mut usize, ) -> Result<Self, SerializationError>
where Self: Sized,

Decode a serializable from a bytestream

Provided Methods

fn verify_fixed( stream: &Vec<u8>, pos: &mut usize, ) -> Result<(), SerializationError>

Verify that the serialized representation of the struct has the correct size, including header + footer.

Will panic for variable sized structs.

fn from_tofpacket(packet: &TofPacket) -> Result<Self, SerializationError>
where Self: Sized,

Decode a serializable directly from a TofPacket

fn to_bytestream(&self) -> Vec<u8>

Encode a serializable to a bytestream

fn from_slice( _slice: &[u8], _start_pos: usize, ) -> Result<Self, SerializationError>
where Self: Sized,

fn to_slice(&self) -> &[u8]
where Self: Sized,

Construct byte slice out of self.

Can not fail.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.