Trait tof_dataclasses::series::MoniSeries

pub trait MoniSeries<T>
where
    T: Copy + MoniData,
{
    // Required methods
    fn get_data(&self) -> &HashMap<u8, VecDeque<T>>;
    fn get_data_mut(&mut self) -> &mut HashMap<u8, VecDeque<T>>;
    fn get_max_size(&self) -> usize;

    // Provided methods
    fn get_var(&self, varname: &str) -> HashMap<u8, Vec<f32>> { ... }
    fn get_var_for_board(&self, varname: &str, rb_id: &u8) -> Option<Vec<f32>> { ... }
    fn keys() -> Vec<&'static str> { ... }
    fn get_board_ids(&self) -> Vec<u8> { ... }
    fn add(&mut self, data: T) { ... }
    fn get_last_moni(&self, board_id: u8) -> Option<T> { ... }
}

A MoniSeries is a collection of (primarily) monitoring data, which comes from multiple senders. E.g. a MoniSeries could hold RBMoniData from all 40 ReadoutBoards.

Required Methods

fn get_data(&self) -> &HashMap<u8, VecDeque<T>>

fn get_data_mut(&mut self) -> &mut HashMap<u8, VecDeque<T>>

fn get_max_size(&self) -> usize

Provided Methods

fn get_var(&self, varname: &str) -> HashMap<u8, Vec<f32>>

A HashMap of <rbid, Vec>

fn get_var_for_board(&self, varname: &str, rb_id: &u8) -> Option<Vec<f32>>

Get a certain variable, but only for a single board

fn keys() -> Vec<&'static str>

A list of the variables in this MoniSeries

fn get_board_ids(&self) -> Vec<u8>

A list of boards in this series

fn add(&mut self, data: T)

Add another instance of the data container to the series

fn get_last_moni(&self, board_id: u8) -> Option<T>

Object Safety

This trait is not object safe.